+++ docker/tests/test_canvas_blocks_footer.py
@@ -23,7 +23,7 @@ class TestCanvasFooterBlocks:
         footer = format_canvas_footer(version="0.8.4", quilt_host="test.com", bucket="test-bucket")
 
         assert isinstance(footer, str)
-        assert "Package will be created/updated asynchronously" in footer
+        assert "Package updates automatically" in footer
         assert "test-bucket@test.com" in footer
         assert "Rev 0.8.4" in footer
 
@@ -61,7 +61,7 @@ class TestCanvasFooterBlocks:
 
         # Verify footer content
         footer_content = last_block.value
-        assert "Package will be created/updated asynchronously" in footer_content
+        assert "Package updates automatically" in footer_content
         assert mock_config.s3_bucket_name in footer_content
         assert mock_config.quilt_catalog in footer_content
 
@@ -99,7 +99,7 @@ class TestCanvasFooterBlocks:
         footer_block_dict = blocks_dict[-1]
         assert footer_block_dict["type"] == "MARKDOWN"
         assert footer_block_dict["id"] == "md-footer"
-        assert "Package will be created/updated asynchronously" in footer_block_dict["value"]
+        assert "Package updates automatically" in footer_block_dict["value"]
 
     def test_blocks_to_dict_rejects_invalid_blocks(self):
         """Verify blocks_to_dict raises TypeError for invalid block types."""